Moved shrinkers into a separate file. Also defined a test for Map of double.
