+++ src/FuncSharp.Tests/Generative/Shrinkers.cs
@@ -1,0 +1,33 @@
+﻿using System.Collections.Generic;
+
+namespace FuncSharp.Tests.Generative;
+
+/// <summary>
+/// Shrinkers are used for lowering the value generated by generators. So that if a test fails, the test framework will try reducing the value to find the lowest failing value.
+/// </summary>
+internal static class Shrinkers
+{
+    internal static IEnumerable<IOption<int>> IntOption(IOption<int> option)
+    {
+        if(option.NonEmpty)
+            yield return Option.Valued(option.GetOrDefault() - 1);
+    }
+
+    internal static IEnumerable<IOption<decimal>> DecimalOption(IOption<decimal> option)
+    {
+        if(option.NonEmpty)
+            yield return Option.Valued(option.GetOrDefault() - 0.1m);
+    }
+
+    internal static IEnumerable<IOption<double>> DoubleOption(IOption<double> option)
+    {
+        if(option.NonEmpty)
+            yield return Option.Valued(option.GetOrDefault() - 0.1d);
+    }
+
+    internal static IEnumerable<IOption<ReferenceType>> ReferenceTypeOption(IOption<ReferenceType> option)
+    {
+        if(option.NonEmpty)
+            yield return Option.Valued(new ReferenceType(option.GetOrDefault().Value - 1));
+    }
+}
